Refactored hard chain logic into common chain logic
